Changed key mapping functions and updated the README
I have had a big issue with the way tslime.vim did it's keybindings.  I use
`ctrl-c` just about every time I exit insert mode, and whenever I want
to leave a command.  By mapping `ctrl-c' the way it did, it made it too
hard for the user to remap the keys.  Hopefully this fix will help with
that.

Changes
-------
* Removed default keymappings for the functions in plugin/tslime.vim and
  replaced them with `<Plug>` mappings that should be less obstrusive
* Added info about the new changes to the keymappings